bug fixed: collision detection LaserShot -> Enemy fixed the issue that some hits don't trigger the collision_detected flag
Problem was in the point_within_area(...) function defined in Math_Functions.cpp. Data was lost due to type conversion from int to long double.
Now: type conversion with static_cast<long double> before division

Furthermore changed value in the if statement from double to long, so the collision detection is more preceise

+++ Math_Functions.cpp
@@ -77,18 +77,18 @@ bool point_within_area(int& P_x, int& P_y, int& A_x, int& A_y, int& B_x, int& B_
 	if (B_x == A_x || D_y == A_y) {
 		return false;
 	}
-	double a = (P_x - A_x) / (B_x - A_x);
-	double b = (D_x - A_x) / (B_x - A_x);
-	double c = (P_y - A_y) / (D_y - A_y);
-	double d = (B_y - A_y) / (D_y - A_y);
+	long double a = static_cast<long double>(P_x - A_x) / (B_x - A_x);
+	long double b = static_cast<long double>(D_x - A_x) / (B_x - A_x);
+	long double c = static_cast<long double>(P_y - A_y) / (D_y - A_y);
+	long double d = static_cast<long double>(B_y - A_y) / (D_y - A_y);
 
-	double s = (c - a * d) / (1 - b * d);
-	double r = a - b * (c - a * d) / (1 - b * d);
+	long double s = (c - a * d) / (1 - b * d);
+	long double r = a - b * (c - a * d) / (1 - b * d);
 
 	
 
-	if (r > 0.0 && r < 1.0 && s > 0.0 && s < 1.0) {
-		std::cout << "s = " << s << "| r = " << r << std::endl;
+	if (r > 0.0l && r < 1.0l && s > 0.0l && s < 1.0l) {
+ 		std::cout << "s = " << s << "| r = " << r << std::endl;
 		return true;
 	}
 	else {
